Mix in "stuck" values into entropy pool
When a raw entropy value of a time delta is identified to be stuck,
i.e. the stuck health test fails on it, it is still mixed into the
entropy pool. In this case, the time delta value is considered
"supplemental data" which is not credited with any entropy.

The Jitter RNG still will generate the required amount of "non-stuck"
time deltas for generating an output buffer: (256 + 64) * osr.

Signed-off-by: Stephan Mueller <smueller@chronox.de>