SHA3-256 replaced by XDRBG
The deterministic output of a random number used to be generated by
simply returning the SHA3-256 digest to the caller.

This behavior now is replaced by using an XDRBG-256 based on SHAKE-256
as defined in [1]. This DRBG is peer-reviewed and planned to be added
to the next SP800-90A revision.

The XDRBG-256 is operated as a conditioning function that only generates
256 bits of output after receiving 256 bits of entropy. It cannot be
operated as a "free-running" DRBG.

This implies, however, that until addition to the 90A the XDRBG-256
remains as a non-vetted conditioning component due to the use of
SHAKE-256. Therefore, in FIPS mode, a SHA3-512-based XDRBG-like
conditioning component is used. Its implementation is identical to
the XDRBG-256 but uses SHA3-512 instead. Due to the use of SHA3-512
there is a subtle state-management difference to the XDRBG-256
which is marked in the code.

The self-test vector for XDRBG-256 is generated using the
implementation from leancrypto.org and covers the initial seeding
and reseeding operations.

The self-test vector for the SHA3-512-based XDRBG-like algorithm
is derived by this implementation.

Finally, the change also updates the hash loop to use SHA3-512
instead of SHA3-256. This is considered to be editorial in nature
because we are only interested in a Keccak operation and not
in the strength of the input/output of that hash.

[1] https://leancrypto.org/papers/xdrbg.pdf

Signed-off-by: Stephan Mueller <smueller@chronox.de>