Use XDRBG-256 for all use cases
NIST confirmed that SHAKE-256 as used in the Jitter RNG can be considered
a vetted conditioning component. Therefore, we can drop the SHA3-512-based
XDRBG-like construction.

To limit the changes to older versions of the Jitter RNG, use SHA3-256
for the hash loop (just like in older versions). Althogh this is window
dressing, because the hash loop operation's purpose is to trigger one
Keccak operation per loop iteration and thus the use of SHA3-512 or
SHA3-256 is immaterial, switch back to reduce the amount of changes.

Signed-off-by: Stephan Mueller <smueller@chronox.de>