FIPS mode: power up phase identical to NTG.1
During the entropy working group discussion with NIST on Sep 2. 2025,
the NIST team suggested that the NTG.1 power up phase where the hash
collection loop and memory access loop is invoked independently is
also used for the FIPS mode.

This approach is considered to be appropriate even without adding an
additional vetting of both mechanisms separately from the commmon
behavior, because the data collection from those sources could be
treated with having zero bits of entropy and yet be SP800-90B
compliant.

As this proposal was made directly by the NIST team, it is therefore
applied before the relase of 3.7.0.

Signed-off-by: Stephan Mueller <smueller@chronox.de>