Time delta calculation: use simple difference
For the calculation of the 1st and 2n discrete derivation of the time
delta (i.e. delta of deltas and delta of deltas of deltas) for the
health test, the goal is to verify that variations exist. This
information is collected by the RCT. With that in mind, it becomes
clear that the direction in which the changes go are important
for the health test as well. This implies that the plus or minus
sign of the actual delta value is important as well.

Therefore, replace the calculation of the absolute value of the
delta with the signed delta.

Signed-off-by: Stephan Mueller <smueller@chronox.de>